changed tmp file removal for stability

+++ src/jansnekokernel/kernel.py
@@ -2,6 +2,8 @@
 from ipykernel.kernelbase import Kernel
 import pexpect, os, shutil
 
+workingdir = "/tmp/nekocode/"
+
 class jansnekokernel(Kernel):
     implementation = 'IPython'
     implementation_version = '8.10.0'
@@ -17,13 +19,14 @@ class jansnekokernel(Kernel):
     def do_execute(self, code, silent, store_history=True, user_expressions=None,
                    allow_stdin=False):
         if not silent:            
-            workingdir = "/tmp/nekocode/"
+            if os.path.exists(workingdir):
+                shutil.rmtree(workingdir)
             os.mkdir(workingdir)
             with open(workingdir + "project.neko", "w") as f:
                     f.write(code)
-            pexpect.run('nekoc ' + workingdir  + 'project.neko')
-            solution = pexpect.run('neko ' + workingdir + 'project.n').decode('ascii')
-            shutil.rmtree(workingdir)
+            solution = pexpect.run('nekoc ' + workingdir  + 'project.neko')
+            if os.path.exists(workingdir + 'project.n'):
+                solution = pexpect.run('neko ' + workingdir + 'project.n').decode('ascii')
             stream_content = {'name': 'stdout', 'text': solution}
             self.send_response(self.iopub_socket, 'stream', stream_content)
 
@@ -32,3 +35,6 @@ class jansnekokernel(Kernel):
                 'payload': [],
                 'user_expressions': {},
                }
+    
+    def do_shutdown(self, restart):
+        shutil.rmtree(workingdir)
